Add the app controller
